Locate workspace inside openfl dir
